==> 11-data.qmd <==
# Data Sources
Below you will find a list of resources that you might want to explore when sourcing data for your coursework assignment or your dissertation. This is by no means an exhaustive list, but simply contains some suggestions of websites that you may want to use.

::: { .callout-note}
You are **not limited** to using these datasets for your coursework assignment or your dissertation, but these are merely some suggestions .
:::

## Open Data
The following websites contain Open Data or link to Open Data from several respectable data providers:

- [AfricanUrbanNetwork](https://github.com/rafaelprietocuriel/AfricanUrbanNetwork)
- [AirBnB Data](http://insideairbnb.com/get-the-data.html)
- [Bike Docking Data (ready for R)](https://docs.ropensci.org/bikedata/)
- [Bing Maps worldwide road detections](https://github.com/microsoft/RoadDetections)
- [Camden Air Action](https://camdenairaction.wordpress.com/2017/02/20/schools-monitoring-project-spring-2017/)
- [Consumer Data Research Centre](https://data.cdrc.ac.uk/)
- [Department for Environment, Food & Rural Affairs](https://environment.data.gov.uk/)
- [Edina (e.g. OS mastermap)](https://digimap.edina.ac.uk/)
- [EU Tourism Data](https://ec.europa.eu/eurostat/statistics-explained/index.php/Tourism_statistics)
- [Eurostat](https://ec.europa.eu/eurostat)
- [Geofabrik (OSM data)](https://www.geofabrik.de/)
- [Global Urban Areas dataset](https://daylightmap.org/2023/11/27/urban.html)
- [Global Weather Data](https://rp5.ru/Weather_in_the_world)
- [Google Dataset Search](https://datasetsearch.research.google.com/)
- [Google Open Buildings](https://sites.research.google/open-buildings/)
- [Kaggle Public Datasets](https://www.kaggle.com/datasets)
- [King's College Data on Air Pollution](https://www.londonair.org.uk/LondonAir/Default.aspx)
- [London Data Store](https://data.london.gov.uk/)
- [London Local Authority Maintained Trees](https://data.london.gov.uk/dataset/local-authority-maintained-trees)
- [London Tube PM2.5 Levels](https://www.ft.com/content/6f381ad4-fef7-11e9-be59-e49b2a136b8d)
- [Microsoft Global Building Footprints](https://github.com/microsoft/GlobalMLBuildingFootprints#start-of-content)
- [Microsoft Research Open Data](https://msropendata.com/)
- [National Public Transport Access Nodes (NaPTAN)](https://www.data.gov.uk/dataset/ff93ffc1-6656-47d8-9155-85ea0b8f2251/national-public-transport-access-nodes-naptan)
- [NASA EARTHDATA](https://earthdata.nasa.gov/)
- [NASA SocioEconomic Data and Applications Center (SEDAC)](https://sedac.ciesin.columbia.edu/)
- [NHS Data (ready for R)](https://nhs-r-community.github.io/NHSRdatasets/)
- [nomis Official Census and Labour Market Statistics](https://www.nomisweb.co.uk/)
- [Office for National Statistics Geoportal](https://geoportal.statistics.gov.uk/)
- [Office for National Statistics](https://www.ons.gov.uk/)
- [Open Topography](https://opentopography.org/)
- [Planetary Computer Data Catalog](https://planetarycomputer.microsoft.com/catalog)
- [pseudo Census Output Areas 2001-2011-2021](https://profrichharris.github.io/census/harmonised/)
- [Public transport accessibility indicators Great Britain](https://www.nature.com/articles/s41597-023-02890-w)
- [Tesco Store Data (London)](https://www.nature.com/articles/s41597-020-0397-7)
- [TfL Cycling Data](https://cycling.data.tfl.gov.uk/)
- [TfL Open Data](https://tfl.gov.uk/info-for/open-data-users/our-open-data?intcmp=3671#on-this-page-2)
- [Tidy Tuesday Data (not exclusively spatial data)](https://github.com/rfordatascience/tidytuesday)
- [UK Data Service](https://ukdataservice.ac.uk/)
- [US Census Data](https://www.census.gov/data.html)
- [US City Open Data Census](http://us-cities.survey.okfn.org/)
- [USGS Earth Explorer](https://earthexplorer.usgs.gov/)
- [UTD19 Multi-City Traffic Dataset](https://utd19.ethz.ch/)
- [WorldPop GitHub](https://github.com/wpgp)
- [WorldPop](https://www.worldpop.org/)

Some other websites that could be helpful:

- [Awesome Public Datasets](https://github.com/awesomedata/awesome-public-datasets); general collection of datasets, although not limited to spatial data.
- [Free GIS data](https://freegisdata.rtwilson.com/); long list with lots of GIS datasets on many different topics and covering many different areas.

## Safeguarded Data
Undergraduate students can also apply for a **Safeguarded** dataset held by the [Consumer Data Research Centre](https://www.cdrc.ac.uk/). There is a process to access these **Safeguarded** datasets, which is detailed on the [CDRC website](https://data.cdrc.ac.uk/using-our-data-services). Please be aware that it normally takes 4-5 weeks for your application to be processed.

As part of the process, you will need to state in your application why you want that specific dataset and what you are planning to do with it. You will also need to have at least thought about the ethical implications of using that data and provide this with your data application.

Some of the datasets held by the CDRC that you can apply for are: 

- [Bicycle Sharing System Docking Station Observations](https://data.cdrc.ac.uk/dataset/bicycle-sharing-system-docking-station-observations)
- [CDRC Modelled Ethnicity Proportions - LSOA Geography]( https://data.cdrc.ac.uk/dataset/cdrc-modelled-ethnicity-proportions-lsoa-geography)
- [FCA Financial Lives Survey](https://data.cdrc.ac.uk/dataset/fca-financial-lives-survey)
- [Speedchecker Broadband Internet Speed Tests](https://data.cdrc.ac.uk/dataset/speedchecker-broadband-internet-speed-tests)

::: { .callout-tip}
Given that the application can take several weeks, the Safeguarded CDRC datasets may be useful for your undergraduate dissertation but probably not for the GEOG0030 coursework assignment. However, any of the CDRC datasets that are marked as **Open Data** do not require this application process and you can download these datasets directly after registering on the website.
:::
